# my-app/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "Shopping Service API Specialized in coffee",
    "version": "1.0",
    "title": "bean cafe API",
    "termsOfService": "about:blank",
    "contact": {
      "name": "Hongrok Lim",
      "url": "https://github.com/rokong/",
      "email": "hongrr123@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8080",
  "basePath": "/bean-cafe",
...
  "paths": {
...
    "/category/{id}/order": {
      "put": {
        "tags": [
          "Category"
        ],
        "summary": "update category's order",
        "description": "change category's order in siblings [id, ord]",
        "operationId": "updateCategoryOrderUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "category",
            "description": "category",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Category"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/order": {
      "post": {
        "tags": [
          "Order"
        ],
        "summary": "create order",
        "description": "create basic order information",
        "operationId": "createOrderUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "order",
            "description": "order",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/order/{id}": {
      "get": {
        "tags": [
          "Order"
        ],
        "summary": "get order",
        "description": "get order by order id [id]",
        "operationId": "getOrderUsingGET",
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/order/{id}/pay": {
      "put": {
        "tags": [
          "Order"
        ],
        "summary": "update order pay",
        "description": "update order pay [id, payId]",
        "operationId": "updateOrderPayUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "order",
            "description": "order",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/order/{id}/status": {
      "put": {
        "tags": [
          "Order"
        ],
        "summary": "update order status",
        "description": "update order status [id, orderStatus]",
        "operationId": "updateOrderStatusUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "order",
            "description": "order",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
...
    "Category": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32",
          "example": 3,
          "description": "category id"
        },
        "name": {
          "type": "string",
          "example": "africa",
          "description": "category name"
        },
        "upId": {
          "type": "integer",
          "format": "int32",
          "example": 7,
          "description": "parent's id"
        },
        "ord": {
          "type": "integer",
          "format": "int32",
          "example": 2,
          "description": "display order in siblings",
          "minimum": 1,
          "exclusiveMinimum": false,
          "exclusiveMaximum": false
        }
      },
      "title": "Category",
      "description": "products are classified in this category"
    },
...
    "Order": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32",
          "example": 1,
          "description": "order id"
        },
        "userNm": {
          "type": "string",
          "example": "customer1",
          "description": "customer name"
        },
        "price": {
          "type": "integer",
          "format": "int32",
          "example": 13700,
          "description": "products' price"
        },
        "deliveryPrice": {
          "type": "integer",
          "format": "int32",
          "example": 5500,
          "description": "delivery price"
        },
        "payId": {
          "type": "integer",
          "format": "int32",
          "example": 7,
          "description": "pay id"
        },
        "payNm": {
          "type": "string",
          "description": "pay name"
        },
        "requestDt": {
          "type": "string",
          "example": "2020-10-18",
          "description": "order request date"
        },
        "lastEditDt": {
          "type": "string",
          "example": "2020-10-18",
          "description": "last changed date"
        },
        "orderStatus": {
          "type": "string",
          "example": "WRITING",
          "description": "order status",
          "enum": [
            "WRITING",
            "PAYMENT",
            "PAYMENT_STANDBY",
            "PAYMENT_PROGRESS",
            "CHECKING",
            "CHECK_COMPLETE",
            "PRODUCT_READY",
            "DELIVERING",
            "COMPLETE",
            "CANCEL",
            "CUSTOMER_CANCEL",
            "CANCELED_WRITE",
            "CANCELED_PAYMENT",
            "CANCELED_CHECK",
            "CANCELED_PRODUCT",
            "SELLER_CANCEL",
            "CANCEL_CHECK",
            "CANCEL_PRODUCT",
            "CANCEL_DELIVER"
          ]
        }
      },
      "title": "Order",
      "description": "Order including primary information"
    },
...
  }
}